# contracts/openapi-v0001.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Value Investment Platform API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/research/runs": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Run",
        "operationId": "run_api_research_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 240,
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "research"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_research_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Api Research Runs Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/research/runs/{run_id}": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Read",
        "operationId": "read_api_research_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/demo/runs": {
      "get": {
        "tags": [
          "demo"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_demo_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Create Run",
        "description": "Authorized business write (data_admin). Writes run + audit + outbox in ONE txn.",
        "operationId": "create_run_api_demo_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/demo/outbox/{outbox_id}/dispatch": {
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Mark Dispatched",
        "operationId": "mark_dispatched_api_demo_outbox__outbox_id__dispatch_post",
        "parameters": [
          {
            "name": "outbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outbox Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/watchlist/{security_id}": {
      "post": {
        "tags": [
          "collab"
        ],
        "summary": "Add Watch",
        "operationId": "add_watch_api_me_watchlist__security_id__post",
        "parameters": [
          {
            "name": "security_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Security Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/notes": {
      "get": {
        "tags": [
          "collab"
        ],
        "summary": "List Notes",
        "operationId": "list_notes_api_me_notes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "collab"
        ],
        "summary": "Add Note",
        "operationId": "add_note_api_me_notes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/worker/outbox/{outbox_id}/claim": {
      "post": {
        "tags": [
          "worker"
        ],
        "summary": "Claim Outbox",
        "operationId": "claim_outbox_api_worker_outbox__outbox_id__claim_post",
        "parameters": [
          {
            "name": "outbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outbox Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/worker/outbox/{outbox_id}/retry": {
      "post": {
        "tags": [
          "worker"
        ],
        "summary": "Retry",
        "operationId": "retry_api_worker_outbox__outbox_id__retry_post",
        "parameters": [
          {
            "name": "outbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outbox Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NoteIn": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "body"
        ],
        "title": "NoteIn"
      },
...
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "partial",
              "completed"
            ],
            "title": "Status"
          },
          "manifest_hash": {
            "type": "string",
            "title": "Manifest Hash"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "manifest": {
            "additionalProperties": true,
            "type": "object",
            "title": "Manifest"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "manifest_hash",
          "created_at",
          "manifest",
          "result"
        ],
        "title": "RunDetail"
      },
      "RunIn": {
        "properties": {
          "source_key": {
            "type": "string",
            "title": "Source Key"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "source_key"
        ],
        "title": "RunIn"
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "partial",
              "completed"
            ],
            "title": "Status"
          },
          "manifest_hash": {
            "type": "string",
            "title": "Manifest Hash"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "manifest_hash",
          "created_at"
        ],
        "title": "RunSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}